+++ Task_4/Task_4.cpp
@@ -2,7 +2,7 @@
 
 int main() {
 
-    long long int X = 0, Y = 0;
+    int X = 0, Y = 0;
     std::cout << "Введите X, Y " << std::endl;
     std::cin >> X >> Y;
     
@@ -15,7 +15,7 @@ int main() {
         Y = 0;
     }
     
-    long double A = 0, B = 0, C = 0, K = 0;
+    double A = 0, B = 0, C = 0, K = 0;
     std::cout << "Введите A, B, C, K " << std::endl;
     std::cin >> A >> B >> C >> K;
